#2258. Rename and reorder type promotion tests
