+++ LanguageFeatures/Private-fields-promotion/not_promotable_A02_t02.dart
@@ -12,37 +12,26 @@
 /// - There is no implicit noSuchMethod forwarder with the same name elsewhere
 ///   in the library.
 ///
-/// @description Checks that if there are other concrete instance getters with
-/// the same name in the same library and they are not a final fields then the
-/// field is not promotable. Test the case when there is a getter with the same
-/// name in some extension in the same library. In this case promotion is
-/// allowed
+/// @description Checks that an instance field is not promotable if it is not
+/// private. Test type variable
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=inference-update-2
 
-class A {}
-
 class C<T> {
-  final T _x;
-  C(this._x);
+  final T x;
+  C(this.x);
 
-  void testC() {
-    if (_x is int) {
-      _x.isOdd;
+  void test() {
+    if (x is int) {
+      x.isOdd;
+//      ^^^^^
+// [analyzer] unspecified
+// [cfe] unspecified
     }
   }
 }
 
-extension on A {
-  String get _x => "Lily was here";
-}
-
-extension on C {
-  int get _x => 42;
-}
-
 main() {
-  A();
-  C(43).testC();
+  C(42).test();
 }
