+++ LanguageFeatures/Private-fields-promotion/promotion_A04_t01.dart
@@ -12,33 +12,37 @@
 /// - There is no implicit noSuchMethod forwarder with the same name elsewhere
 ///   in the library.
 ///
-/// @description Checks that an instance field is not promotable if it is not
-/// final
+/// @description Checks that if there are other concrete instance getters with
+/// the same name in the same library and they are not a final fields then the
+/// field is not promotable. Test the case when there is a getter with the same
+/// name in some extension in the same library. In this case promotion is
+/// allowed
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=inference-update-2
 
-class C {
-  int? _x;
-  int? _y = 42;
+class A {}
+
+class C<T> {
+  final T _x;
   C(this._x);
 
-  void test() {
-    if (_x != null) {
+  void testC() {
+    if (_x is int) {
       _x.isOdd;
-//       ^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
-    }
-    if (_y != null) {
-      _y.isOdd;
-//       ^^^^^
-// [analyzer] unspecified
-// [cfe] unspecified
     }
   }
 }
 
+extension on A {
+  String get _x => "Lily was here";
+}
+
+extension on C {
+  int get _x => 42;
+}
+
 main() {
-  C(42).test();
+  A();
+  C(43).testC();
 }
